requirements.txt
Django==1.10
django-extensions==1.6.7
djangorestframework==3.4.1
mysqlclient==1.3.7
